+++ vertical-pod-autoscaler/admission-controller/main.go
@@ -49,9 +49,9 @@ func main() {
 	certs := initCerts(*certsDir)
 	stopChannel := make(chan struct{})
 	vpaLister := newReadyVPALister(stopChannel)
-	as := &admissionServer{logic.NewRecommendationProvider(vpaLister, vpa_api_util.NewCappingRecommendationProcessor())}
+	as := logic.NewAdmissionServer(logic.NewRecommendationProvider(vpaLister, vpa_api_util.NewCappingRecommendationProcessor()))
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		as.serve(w, r)
+		as.Serve(w, r)
 	})
 	clientset := getClient()
 	server := &http.Server{
