+++ vertical-pod-autoscaler/updater/main.go
@@ -23,6 +23,7 @@ import (
 	"k8s.io/autoscaler/vertical-pod-autoscaler/common"
 	vpa_clientset "k8s.io/autoscaler/vertical-pod-autoscaler/pkg/client/clientset/versioned"
 	vpa_api_util "k8s.io/autoscaler/vertical-pod-autoscaler/pkg/utils/vpa"
+	updater "k8s.io/autoscaler/vertical-pod-autoscaler/updater/logic"
 	kube_client "k8s.io/client-go/kubernetes"
 	kube_restclient "k8s.io/client-go/rest"
 	"time"
@@ -46,7 +47,7 @@ func main() {
 	// TODO monitoring
 
 	kubeClient, vpaClient := createKubeClients()
-	updater := NewUpdater(kubeClient, vpaClient, *minReplicas, *evictionToleranceFraction, vpa_api_util.NewCappingRecommendationProcessor())
+	updater := updater.NewUpdater(kubeClient, vpaClient, *minReplicas, *evictionToleranceFraction, vpa_api_util.NewCappingRecommendationProcessor())
 	for {
 		select {
 		case <-time.After(*updaterInterval):
